tower-teachers/pyproject.toml
[project]
name = "tower-teachers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "specklepy>=3.2.3",
    "python-dotenv>=1.0.0",
]
